safari-534.53-branch/LayoutTests/fast/viewport/viewport-118-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "" not recognized and ignored.
viewport size 320x352 scale 1.000000 with limits [1.000000, 5.000000] and userScalable -1.000000
